fix(profileactions): reduced considereably the amount of localstorage code in profileactions
